fixed test
