--- cypress/e2e/consortia/inventory/instance/duplicating-local-instance-on-member-with-source-marc.cy.js
@@ -1,88 +1,0 @@
-import { DEFAULT_JOB_PROFILE_NAMES, INSTANCE_SOURCE_NAMES } from '../../../../support/constants';
-import Affiliations, { tenantNames } from '../../../../support/dictionary/affiliations';
-import Permissions from '../../../../support/dictionary/permissions';
-import DataImport from '../../../../support/fragments/data_import/dataImport';
-import InstanceRecordView from '../../../../support/fragments/inventory/instanceRecordView';
-import InventoryInstance from '../../../../support/fragments/inventory/inventoryInstance';
-import InventoryInstances from '../../../../support/fragments/inventory/inventoryInstances';
-import InventoryNewInstance from '../../../../support/fragments/inventory/inventoryNewInstance';
-import ConsortiumManager from '../../../../support/fragments/settings/consortium-manager/consortium-manager';
-import TopMenu from '../../../../support/fragments/topMenu';
-import Users from '../../../../support/fragments/users/users';
-import getRandomPostfix from '../../../../support/utils/stringTools';
-
-describe('Inventory', () => {
-  describe('Instance', () => {
-    const marcFile = {
-      marc: 'oneMarcBib.mrc',
-      marcFileName: `C410924 marcFileName${getRandomPostfix()}.mrc`,
-    };
-    const testData = {
-      newResourceTitle: `C410924 instanceTitle${getRandomPostfix()}`,
-      newResourceType: 'notated movement',
-      source: INSTANCE_SOURCE_NAMES.FOLIO,
-    };
-
-    before('Create test data and login', () => {
-      cy.getAdminToken();
-      cy.setTenant(Affiliations.College);
-      DataImport.uploadFileViaApi(
-        marcFile.marc,
-        marcFile.marcFileName,
-        DEFAULT_JOB_PROFILE_NAMES.CREATE_INSTANCE_AND_SRS,
-      ).then((response) => {
-        testData.instanceId = response[0].instance.id;
-      });
-
-      cy.resetTenant();
-      cy.createTempUser([Permissions.inventoryAll.gui]).then((userProperties) => {
-        testData.user = userProperties;
-
-        cy.assignAffiliationToUser(Affiliations.College, testData.user.userId);
-        cy.setTenant(Affiliations.College);
-        cy.assignPermissionsToExistingUser(testData.user.userId, [Permissions.inventoryAll.gui]);
-
-        cy.login(testData.user.username, testData.user.password, {
-          path: TopMenu.inventoryPath,
-          waiter: InventoryInstances.waitContentLoading,
-        });
-        ConsortiumManager.switchActiveAffiliation(tenantNames.central, tenantNames.college);
-        ConsortiumManager.checkCurrentTenantInTopMenu(tenantNames.college);
-        InventoryInstances.searchByTitle(testData.instanceId);
-        InventoryInstances.selectInstance();
-        InventoryInstance.waitLoading();
-      });
-    });
-
-    after('Delete test data', () => {
-      cy.resetTenant();
-      cy.getAdminToken();
-      Users.deleteViaApi(testData.user.userId);
-      cy.setTenant(Affiliations.College);
-      InventoryInstance.deleteInstanceViaApi(testData.instanceId);
-      cy.getInstance({
-        limit: 1,
-        expandAll: true,
-        query: `"hrid"=="${testData.instanceHrid}"`,
-      }).then((instance) => {
-        InventoryInstance.deleteInstanceViaApi(instance.id);
-      });
-    });
-
-    it(
-      'C410924 (CONSORTIA) Duplicating local instance on Member tenant with Source MARC (folijet)',
-      { tags: ['extendedPathECS', 'folijet'] },
-      () => {
-        InstanceRecordView.duplicate();
-        InventoryNewInstance.fillResourceTitle(testData.newResourceTitle);
-        InventoryNewInstance.fillResourceType(testData.newResourceType);
-        InventoryNewInstance.clickSaveAndCloseButton();
-        InventoryInstance.waitInstanceRecordViewOpened(testData.newResourceTitle);
-        InventoryInstance.checkInstanceDetails([{ key: 'Source', value: testData.source }]);
-        InventoryInstance.getAssignedHRID().then((initialInstanceHrId) => {
-          testData.instanceHrid = initialInstanceHrId;
-        });
-      },
-    );
-  });
-});
